Updated api endpoints for JWT

+++ src/main/java/com/edmond/bank/api/UserRestController.java
@@ -10,46 +10,54 @@ import org.springframework.security.oauth2.server.resource.authentication.JwtAut
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@CrossOrigin
+@CrossOrigin("*")
 @RequestMapping("api/user")
 public class UserRestController {
 
     @Autowired
     UserService userService;
 
-//    @GetMapping("/{userId}")
-//    public User getUser(@PathVariable int userId, JwtAuthenticationToken auth) {
-//        String email = (String) auth.getToken().getClaims().get("email");
-//        User searchUser = userService.findById(userId);
-//        if (searchUser.getUsername().equalsIgnoreCase(email))
-//            return userService.findById(userId);
-//        else
-//            throw new UnauthorizedException("You are not authorized");
-//    }
-
-
-    @GetMapping("/{userId}")
-    public ResponseEntity<Object> getUser(@PathVariable int userId, JwtAuthenticationToken auth) {
-        User result = userService.findById(userId); //Use optional
+    @GetMapping
+    public ResponseEntity<Object> getUser(JwtAuthenticationToken auth) {
         String email = (String) auth.getToken().getClaims().get("email");
+        User result = userService.findUserByEmail(email); //Use optional
+
         if (result.getUsername().equalsIgnoreCase(email))
             return ResponseHandler.generateResponse("OK", HttpStatus.OK, result);
         else
             return ResponseHandler.generateResponse("Unauthorized Access", HttpStatus.UNAUTHORIZED, null);
     }
 
+    /**
+     * TODO: This is an endpoint that should only be adminstered by an ADMIN. Put in AccountRestController instead
+     * @param user
+     * @param auth
+     * @return
+     */
     @PostMapping
-    public User addUser(@RequestBody User user) {
+    public User addUser(@RequestBody User user, JwtAuthenticationToken auth) {
         return userService.createUser(user);
     }
     @DeleteMapping("/{userId}")
-    public void deleteUser(@PathVariable int userId) {
-        userService.deleteById(userId);
+    public void deleteUser(@PathVariable int userId, JwtAuthenticationToken auth) {
+        //TODO: Throw an error if not authenticated or authorized inside isUserAuthenticated method
+        if(isUserAuthenticated(userId, auth))
+            userService.deleteById(userId);
     }
 
     @PutMapping("/{userId}")
-    public User updateUser(@RequestBody User user) {
-        this.userService.editUser(user);
+    public User updateUser(@PathVariable int userId, @RequestBody User user, JwtAuthenticationToken auth) {
+        if(isUserAuthenticated(userId, auth))
+            this.userService.editUser(user);
+
+        //TODO: Throw an error if not authenticated or authorized inside isUserAuthenticated method
         return user;
+    }
+
+    private boolean isUserAuthenticated(int userId, JwtAuthenticationToken auth) {
+        String email = (String) auth.getToken().getClaims().get("email");
+        User result = userService.findUserByEmail(email); //Use optional
+
+        return result.getUsername().equalsIgnoreCase(email);
     }
 }
